camps/camp1-identity/secure-server/pyproject.toml
[project]
name = "camp1-secure-server"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "uvicorn>=0.30.0",
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.8.0",
    "pyjwt>=2.8.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
